fixed signed mismatches

+++ bioinformatic algorithms/hw5/problem17_bwt/suffix_array.h
@@ -18,12 +18,12 @@ static void radixPass(std::basic_string<size_t>& a, std::basic_string<size_t> & 
   }
   for (int i = 0;  i < n;  i++) b[c[r[a[i]]]++] = a[i];      // sort
 }
- 
+
 // find the suffix array SA of s[0..n-1] in {1..K}^n
 // require s[n]=s[n+1]=s[n+2]=0, n>=2
-void suffixArray(std::basic_string<size_t> & s, std::basic_string<size_t>& SA, int n, int K) {
-  int n0=(n+2)/3, n1=(n+1)/3, n2=n/3, n02=n0+n2;
-  std::basic_string<size_t> s12(n02 + 3, 0);  
+void suffixArray(std::basic_string<size_t> & s, std::basic_string<size_t>& SA, size_t n, size_t K) {
+  size_t n0=(n+2)/3, n1=(n+1)/3, n2=n/3, n02=n0+n2;
+  std::basic_string<size_t> s12(n02 + 3, 0);
   s12[n02]= s12[n02+1]= s12[n02+2]=0;
   std::basic_string<size_t> SA12(n02 + 3, 0);
   SA12[n02]=SA12[n02+1]=SA12[n02+2]=0;
@@ -32,7 +32,9 @@ void suffixArray(std::basic_string<size_t> & s, std::basic_string<size_t>& SA, i
 
   // generate positions of mod 1 and mod  2 suffixes
   // the "+(n0-n1)" adds a dummy mod 1 suffix if n%3 == 1
-  for (int i=0, j=0;  i < n+(n0-n1);  i++) if (i%3 != 0) s12[j++] = i;
+  for (size_t i = 0, j = 0;  i < n + (n0 - n1);  i++)
+    if (i%3 != 0)
+      s12[j++] = i;
 
   // lsb radix sort the mod 1 and mod 2 triples
   radixPass(s12 , SA12, s.begin() + 2, n02, K);
@@ -40,8 +42,8 @@ void suffixArray(std::basic_string<size_t> & s, std::basic_string<size_t>& SA, i
   radixPass(s12 , SA12, s.begin()  , n02, K);
 
   // find lexicographic names of triples
-  int name = 0, c0 = -1, c1 = -1, c2 = -1;
-  for (int i = 0;  i < n02;  i++) {
+  size_t name = 0, c0 = -1, c1 = -1, c2 = -1;
+  for (size_t i = 0;  i < n02;  i++) {
     if (s[SA12[i]] != c0 || s[SA12[i]+1] != c1 || s[SA12[i]+2] != c2) {
       name++;  c0 = s[SA12[i]];  c1 = s[SA12[i]+1];  c2 = s[SA12[i]+2];
     }
@@ -53,16 +55,17 @@ void suffixArray(std::basic_string<size_t> & s, std::basic_string<size_t>& SA, i
   if (name < n02) {
     suffixArray(s12, SA12, n02, name);
     // store unique names in s12 using the suffix array
-    for (int i = 0;  i < n02;  i++) s12[SA12[i]] = i + 1;
+    for (size_t i = 0;  i < n02;  i++)
+        s12[SA12[i]] = i + 1;
   } else // generate the suffix array of s12 directly
-    for (int i = 0;  i < n02;  i++) SA12[s12[i] - 1] = i;
+    for (size_t i = 0;  i < n02;  i++) SA12[s12[i] - 1] = i;
 
   // stably sort the mod 0 suffixes from SA12 by their first character
-  for (int i=0, j=0;  i < n02;  i++) if (SA12[i] < n0) s0[j++] = 3*SA12[i];
+  for (size_t i=0, j=0;  i < n02;  i++) if (SA12[i] < n0) s0[j++] = 3*SA12[i];
   radixPass(s0, SA0, s.begin(), n0, K);
 
   // merge sorted SA0 suffixes and sorted SA12 suffixes
-  for (int p=0,  t=n0-n1,  k=0;  k < n;  k++) {
+  for (size_t p = 0,  t=n0-n1,  k=0;  k < n;  k++) {
 #define GetI() (SA12[t] < n0 ? SA12[t] * 3 + 1 : (SA12[t] - n0) * 3 + 2)
     int i = GetI(); // pos of current offset 12 suffix
     int j = SA0[p]; // pos of current offset 0  suffix
